Add files via upload
This is to showcase my completed school work. In this Project & Portfolio 1 class I was able to apply my knowledge of OS, networking and security practices in order to plan a technology solutions based on a typical use case. I had to demonstrate and troubleshoot the design, installation and configuration of a working project with servers and user endpoints. Also, I have here to show my written project plan documenting my work.